pyproject.toml
[tool.poetry]
name = "potg"
version = "0.1.0"
description = "POTG - FastAPI Authentication Service"
authors = ["Your Name <your-email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
prisma = "^0.15.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {extras = ["email"], version = "^2.7.0"}
pydantic-settings = "^2.3.0"
python-multipart = "^0.0.9"
aiosmtplib = "^3.0.1"
email-validator = "^2.2.0"
python-dotenv = "^1.0.1"
bcrypt = "^5.0.0"
cloudinary = "^1.44.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry-core.masonry.api"
